# contracts/openapi/inspector-api.v1.yaml
openapi: 3.0.3
info:
  title: Инспектор ИИ — REST API
  version: 0.24.0
  description: |
    Публичный REST API сервиса «Инспектор ИИ» (веб-клиент и внешние ИС, целевая — ИАИС «РиН»).
    Асинхронная pull-модель: загрузка → process_id → опрос статуса → получение протокола.

    Требования и их ID — docs/TZ.md.

    Соглашения:
    - все поля snake_case; идентификаторы — UUID (кроме params.id — integer, как в ТЗ);
    - время — RFC 3339 (UTC);
    - bbox — [x0, y0, x1, y1] в долях [0;1] видимой страницы (после CropBox/Rotate),
      начало координат — левый верхний угол (как viewport pdf.js);
    - страницы нумеруются с 1;
    - ошибки — объект Error.

    Изменения 0.2.0: пары страниц для экрана сравнения (PagePair, GET /protocols/{id}/page-pairs),
    compliance_percent в сводке, auto_start при загрузке, метод гипотез VISUAL_DIFF, фильтры findings по file_id/evidence_page,
    Finding.page_pair_id и rationale_source.

    Изменения 0.3.0: манифест комплекта при загрузке
    (UploadManifest), external_id объекта, серверный импорт папки (POST /documents/import),
    правка доказательств новой версией (POST /findings/{id}/evidence, evidence_history),
    массовое решение CONFIRM/CLARIFY по одному параметру (POST /findings/bulk-decision),
    PATCH /files — только выбор авторитетной редакции (метаданные не редактируются),
    сводка комплектности (ProcessInfo.completeness).

    Изменения 0.4.0 (Приложение 1 и «Перечень ИД» ред. 1.1): коды параметров M-001…M-132;
...

...
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/params:
    get:
      tags: [admin]
      operationId: listParams
      summary: Матрица контроля (132 параметра)
      parameters:
        - name: section
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatrixParam'
    post:
      tags: [admin]
      operationId: createParam
      summary: Добавить параметр (ADMIN) — создаёт новую matrix_version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatrixParamInput'
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatrixParam'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/params/{param_id}:
    parameters:
      - name: param_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [admin]
      operationId: getParam
      summary: Параметр матрицы по id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatrixParam'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [admin]
      operationId: updateParam
      summary: Изменить параметр/пороги без перекодирования (модуль 8) — новая matrix_version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatrixParamInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatrixParam'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [admin]
      operationId: deactivateParam
      summary: Деактивировать параметр (is_active=false)
      responses:
        '204':
          description: Деактивирован
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/admin/normative-docs:
    get:
      tags: [admin]
      operationId: listNormativeDocs
      summary: Нормативная база (модуль 8)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NormativeDoc'
    post:
      tags: [admin]
      operationId: createNormativeDoc
      summary: Добавить нормативный документ (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormativeDocInput'
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormativeDoc'
  /api/v1/admin/normative-docs/{normative_doc_id}:
    parameters:
      - name: normative_doc_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [admin]
      operationId: updateNormativeDoc
      summary: Изменить нормативный документ (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormativeDocInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormativeDoc'
    delete:
      tags: [admin]
      operationId: deactivateNormativeDoc
      summary: Деактивировать нормативный документ (ADMIN)
      responses:
        '204':
          description: Деактивирован (effective_to = сегодня)
  /api/v1/admin/logical-rules:
    get:
      tags: [admin]
      operationId: listLogicalRules
      summary: База логических правил
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogicalRule'
    post:
      tags: [admin]
      operationId: createLogicalRule
      summary: Добавить логическое правило (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogicalRuleInput'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalRule'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/admin/logical-rules/{rule_id}:
    parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [admin]
      operationId: updateLogicalRule
      summary: Изменить логическое правило (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogicalRuleInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalRule'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [admin]
      operationId: deactivateLogicalRule
      summary: Деактивировать логическое правило (ADMIN)
      responses:
        '204':
          description: Деактивировано
  /api/v1/audit:
    get:
      tags: [audit]
      operationId: listAudit
      summary: Журнал аудита (модуль 9)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: actor_type
          in: query
          description: Только действия пользователей (USER) или только системы (SYSTEM)
          schema:
            type: string
            enum: [USER, SYSTEM]
        - name: date_from
          in: query
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPage'
  /api/v1/notifications:
    get:
      tags: [notifications]
      operationId: listNotifications
      summary: Уведомления текущего пользователя
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
  /api/v1/notifications/{notification_id}/read:
    parameters:
      - name: notification_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [notifications]
      operationId: markNotificationRead
      summary: Отметить уведомление прочитанным
      responses:
        '204':
          description: OK
  /api/v1/ml/dataset-items:
    get:
      tags: [ml]
      operationId: listDatasetItems
      summary: Черновик/релизы GOLD-набора (REQ-ML-01..03)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: curation_status
          in: query
          schema:
            $ref: '#/components/schemas/CurationStatus'
        - name: dataset_version
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetItemPage'
  /api/v1/ml/dataset-items/{item_id}/curate:
    parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [ml]
      operationId: curateDatasetItem
      summary: Решение куратора данных по записи GOLD
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mtls:
      type: http
      scheme: bearer
      description: >-
        Для внешних ИС — взаимная TLS-аутентификация клиентским сертификатом (УКЭП, REQ-INT-02).
        OpenAPI 3.0 не имеет типа mutualTLS, поэтому схема описана текстом.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    BadRequest:
      description: Некорректный запрос (валидация OpenAPI)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ReviewPriority:
      type: string
      description: Только очерёдность экспертной проверки, не юридическое действие
      enum: [HIGH, MEDIUM, LOW]
...
    CurationStatus:
      type: string
      enum: [DRAFT, APPROVED, EXCLUDED]
...
    MatrixParamInput:
      type: object
      description: Поля таблицы Params (ТЗ §8.1)
      required: [code, section, parameter_name, data_type, review_priority]
      properties:
        code:
          type: string
          maxLength: 20
          example: M-055
        external_code:
          type: string
          maxLength: 20
          nullable: true
          description: Код параметра у организаторов (каталог parameter_catalog_132) — для выгрузки submission
          example: KR-055
        section:
          type: string
          maxLength: 50
          example: КР
        parameter_name:
          type: string
          maxLength: 255
        unit:
          type: string
          maxLength: 20
          nullable: true
        source_pd:
          type: string
          nullable: true
        source_rd:
          type: string
          nullable: true
        source_id:
          type: string
          nullable: true
          description: Источник в ИД (не путать с идентификатором)
        trigger_logic:
          type: string
          nullable: true
          example: < 0.9 м
        review_priority:
          $ref: '#/components/schemas/ReviewPriority'
        sp_reference:
          type: string
          nullable: true
        gost_reference:
          type: string
          nullable: true
        fz_reference:
          type: string
          nullable: true
        other_normative:
          type: string
          nullable: true
        data_type:
          $ref: '#/components/schemas/ParamDataType'
        min_value:
          type: number
          nullable: true
        max_value:
          type: number
          nullable: true
        regex_pattern:
          type: string
          maxLength: 255
          nullable: true
        semantic_anchors:
          type: array
          description: Расширение — фразы-якоря для семантического поиска
          items:
            type: string
        enum_values:
          type: array
          description: Расширение — упорядоченные значения для data_type=enum (например классы бетона)
          items:
            type: string
        is_active:
          type: boolean
          default: true
    MatrixParam:
      allOf:
        - $ref: '#/components/schemas/MatrixParamInput'
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    NormativeDocInput:
      type: object
      required: [document_name, document_number]
      properties:
        document_name:
          type: string
        document_number:
          type: string
        section:
          type: string
          nullable: true
        parameter_name:
          type: string
          nullable: true
        min_value:
          type: number
          nullable: true
        max_value:
          type: number
          nullable: true
        effective_from:
          type: string
          format: date
          nullable: true
        effective_to:
          type: string
          format: date
          nullable: true
    NormativeDoc:
      allOf:
        - $ref: '#/components/schemas/NormativeDocInput'
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    LogicalRuleInput:
      type: object
      required: [rule_name, condition, expected]
      properties:
        rule_name:
          type: string
        condition:
          type: string
          description: >-
            Условие на извлечённых значениях, язык правил (docs/domain/matrix.md): ссылки
            «КОД.СТАДИЯ», сравнения, and / or / not, exists() и missing(). Например
            «exists(M-002.PD) and exists(M-002.RD)». Проверяется при сохранении: неверное выражение
            или неизвестный код параметра — 400 RULE_INVALID.
        expected:
          type: string
          description: 'Ожидаемое следствие, тот же язык: например «M-002.RD == M-002.PD»'
        normative_base:
          type: string
          nullable: true
        review_priority:
          $ref: '#/components/schemas/ReviewPriority'
        is_active:
          type: boolean
          default: true
    LogicalRule:
      allOf:
        - $ref: '#/components/schemas/LogicalRuleInput'
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid

    # ----------------------------------------------------------------- audit
...
    AuditPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AuditEntry'
    Notification:
      type: object
      required: [id, type, message, created_at, is_read]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - PROTOCOL_READY
            - PROCESS_FAILED
            - NEW_DOCUMENTS_AVAILABLE
            - SYNC_FAILED
            - ADMIN_ALERT
            - RETRAIN_ITEM_PENDING
            - UNFINALIZE_REQUESTED
          description: >-
            RETRAIN_ITEM_PENDING (0.19.0) — отказ инспектора дал запись для дообучения, ждёт решения ADMIN;
            UNFINALIZE_REQUESTED (0.19.0) — инспектор просит откат финализации, адресовано ADMIN и SUPERVISOR
        message:
          type: string
        process_id:
          type: string
          format: uuid
          nullable: true
        object_id:
          type: string
          format: uuid
          nullable: true
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time

    # -------------------------------------------------------------------- ml
...